+++ addOns/quickstart/src/main/java/org/zaproxy/zap/extension/quickstart/ajaxspider/ExtensionQuickStartAjaxSpider.java
@@ -19,30 +19,33 @@
  */
 package org.zaproxy.zap.extension.quickstart.ajaxspider;
 
+import java.net.URI;
 import java.util.List;
 import org.parosproxy.paros.Constant;
 import org.parosproxy.paros.control.Control;
 import org.parosproxy.paros.extension.Extension;
 import org.parosproxy.paros.extension.ExtensionAdaptor;
 import org.parosproxy.paros.extension.ExtensionHook;
-import org.parosproxy.paros.view.View;
-import org.zaproxy.addon.pscan.ExtensionPassiveScan2;
+import org.parosproxy.paros.model.Model;
 import org.zaproxy.zap.extension.quickstart.ExtensionQuickStart;
-import org.zaproxy.zap.extension.selenium.ExtensionSelenium;
+import org.zaproxy.zap.extension.quickstart.ModernSpiderOption;
+import org.zaproxy.zap.extension.spiderAjax.AjaxSpiderParam;
+import org.zaproxy.zap.extension.spiderAjax.AjaxSpiderTarget;
 import org.zaproxy.zap.extension.spiderAjax.ExtensionAjax;
 
 /**
- * Provides the option to use the Ajax Spider when running a quick scan. This is a separate
- * extension so that the main extension still loads if the Ajax Spider is not installed.
+ * Provides the option to use the Ajax Spider as the modern spider when running a quick scan. This
+ * is a separate extension so that the main extension still loads if the Ajax Spider is not
+ * installed.
  */
 public class ExtensionQuickStartAjaxSpider extends ExtensionAdaptor {
 
     public static final String NAME = "ExtensionQuickStartAjaxSpider";
 
     private static final List<Class<? extends Extension>> DEPENDENCIES =
-            List.of(ExtensionAjax.class, ExtensionSelenium.class, ExtensionPassiveScan2.class);
+            List.of(ExtensionQuickStart.class, ExtensionAjax.class);
 
-    private AjaxSpiderExplorer ase;
+    private AjaxSpiderOption ajaxOption;
 
     public ExtensionQuickStartAjaxSpider() {
         super(NAME);
@@ -55,11 +58,9 @@ public class ExtensionQuickStartAjaxSpider extends ExtensionAdaptor {
 
     @Override
     public void hook(ExtensionHook extensionHook) {
-        super.hook(extensionHook);
-
         if (hasView()) {
-            this.ase = new AjaxSpiderExplorer(this);
-            this.getExtQuickStart().addPlugableSpider(ase);
+            this.ajaxOption = new AjaxSpiderOption();
+            getExtQuickStart().addModernSpiderOption(ajaxOption);
         }
     }
 
@@ -71,15 +72,7 @@ public class ExtensionQuickStartAjaxSpider extends ExtensionAdaptor {
     @Override
     public void unload() {
         if (hasView()) {
-            this.getExtQuickStart().removePlugableSpider(ase);
-        }
-    }
-
-    @Override
-    public void optionsLoaded() {
-        super.optionsLoaded();
-        if (View.isInitialised()) {
-            this.ase.optionsLoaded(getExtQuickStart().getQuickStartParam());
+            getExtQuickStart().removeModernSpiderOption(ajaxOption);
         }
     }
 
@@ -102,7 +95,48 @@ public class ExtensionQuickStartAjaxSpider extends ExtensionAdaptor {
         return Control.getSingleton().getExtensionLoader().getExtension(ExtensionQuickStart.class);
     }
 
-    public ExtensionSelenium getExtSelenium() {
-        return Control.getSingleton().getExtensionLoader().getExtension(ExtensionSelenium.class);
+    private class AjaxSpiderOption implements ModernSpiderOption {
+
+        @Override
+        public String getName() {
+            return Constant.messages.getString("quickstart.modern.option.ajaxspider");
+        }
+
+        @Override
+        public String toString() {
+            return getName();
+        }
+
+        @Override
+        public void startScan(URI uri, String browserId) {
+            ExtensionAjax extAjax = getExtAjax();
+            AjaxSpiderParam options =
+                    Model.getSingleton()
+                            .getOptionsParam()
+                            .getParamSet(AjaxSpiderParam.class)
+                            .clone();
+            options.setBrowserId(browserId);
+            AjaxSpiderTarget.Builder builder =
+                    AjaxSpiderTarget.newBuilder(Model.getSingleton().getSession());
+            builder.setStartUri(uri);
+            builder.setInScopeOnly(false);
+            builder.setSubtreeOnly(false);
+            builder.setOptions(options);
+            extAjax.startScan(builder.build());
+        }
+
+        @Override
+        public void stopScan() {
+            getExtAjax().stopScan();
+        }
+
+        @Override
+        public boolean isRunning() {
+            return getExtAjax().isSpiderRunning();
+        }
+
+        private ExtensionAjax getExtAjax() {
+            return Control.getSingleton().getExtensionLoader().getExtension(ExtensionAjax.class);
+        }
     }
 }