+++ addOns/quickstart/CHANGELOG.md
@@ -4,12 +4,12 @@ All notable changes to this add-on will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
-
+### Added
+- Option to select modern spider.
 
 ## [56] - 2026-05-06
 ### Added
 - Add Scan Policy option to the Automated Scan panel.
-
 
 ## [55] - 2026-03-09
 ### Fixed